requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
cron_descriptor==2.0.6
distlib==0.4.0
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-environ==0.12.0
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filelock==3.19.1
flake8==7.3.0
identify==2.6.14
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.14
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
gunicorn==23.0.0
